Use sio2jail on WSL
WSL2 (released 2019) is based on a Linux VM. In general its behavior is
consistent with any other Linux system running in a VM, so there should
be no reason to discriminate WSL under `util.is_linux()`.

In the future, we might want to check for specific features such as
perf counters being supported, but that is an orthogonal concern.

I have tested with several people using WSL across relatively modern
laptops (Intel 8th gen and up) and in all cases the Intel PMU driver
passed through enough perf counters for sio2jail to work.

If we wanted to test support for sio2jail more thoroughly, I suspect the
best method might be to actually try running sio2jail or the perf tool.