Allow compiling LaTeX with lualatex (#282)
This is necessary to use some TeX features, such as TikZ's graph drawing
library.

As there isn't a reasonable way of detecting the necessity of using
lualatex, the --latex-compiler option is now also exposed through
config.yml, to make it possible to export the package if the statement
only compiles with the right latex compiler.